Create folders cosmwasm-std and cosmwasm-testing
