+++ contracts/assemblyscript-poc/contract/src/contract.spec.ts
@@ -1,16 +1,6 @@
 import { query } from "./contract";
-import { Extern } from "./cosmwasm";
+import { makeTestingExtern } from "./cosmwasm-testing";
 import { Encoding } from "./utils";
-
-function makeTestingExtern(): Extern {
-  return new Extern((address: string) => {
-    const addressLength = 20;
-    const encoded = Encoding.toUtf8(address).slice(0, addressLength);
-    const out = new Uint8Array(addressLength);
-    for (let i = 0; i < encoded.length; i++) out[i] = encoded[i];
-    return out;
-  });
-}
 
 describe("contract", () => {
   describe("query", () => {
