+++ contracts/assemblyscript-poc/contract/src/cosmwasm-std/env.ts
