+++ contracts/assemblyscript-poc/contract/src/cosmwasm-std/index.ts
@@ -1,0 +1,1 @@
+export * from "./cosmwasm";
